backend/pyproject.toml
[project]
name = "pm-backend"
version = "0.1.0"
description = "FastAPI backend for the project management MVP"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "itsdangerous>=2.2.0",
    "openai>=2.26.0",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1,<0.29",
    "pytest>=9.0.2",
]
